extract PoC into separate class
